Update NS-004 based on discussion in NSWG
Incorporate feedback from Clint as discussed at CA/BF F2F #62

NB: This had to be rebased onto a branch that also changed the
definition of Certficate Systems, to Certificate System, with a list
similar to the one initially contained in this commit, but in fact more
extensive.

The list after Ballot NS-003 was:

   1. identity validation;
   2. identity authentication;
   3. account registration;
   4. certificate application;
   5. certificate approval;
   6. certificate issuance;
   7. certificate revocation;
   8. authoritative certificate status; or
   9. key escrow.

in the NS-004 pull request, the list was:

* certificate application;
* certificate validation;
* certificate approval;
* signing;
* certificate revocation;
* serving authoritative certificate status information; or
* key escrow.

I have resolved this conflict by taking "certificate validation" and
inserting it into the list as of NS-003 as the fifth item.

    -- tobij