CA Infrastructure Scope update
Following discussions in the NSWG, this is an example version of the NCSSRs which removes several of the "Systems" from the scope of CA Infrastructure (and the document in general).
The changes are based on the assumed completion of IPR review for NS-004 and NS-006.

Additionally, fixed some minor issues arising from NS-004 and NS-006:
* Reversion of Workstation definition
* Orphaned text in 1.2.2
* Markdown formatting in 2.2.1.2