Ballot NS004 - Updating Section 4 - Vulnerability Management - of the NSRs (#34)
* Updating NSR Section 4

This is  the latest PR for the work that the NetSec  team is doing to revamp section 4 of the NSRs

* Update NSR.md

* Update NS-004 based on discussion in NSWG

Incorporate feedback from Clint as discussed at CA/BF F2F #62

NB: This had to be rebased onto a branch that also changed the
definition of Certficate Systems, to Certificate System, with a list
similar to the one initially contained in this commit, but in fact more
extensive.

The list after Ballot NS-003 was:

   1. identity validation;
   2. identity authentication;
   3. account registration;
   4. certificate application;
   5. certificate approval;
   6. certificate issuance;
   7. certificate revocation;
   8. authoritative certificate status; or
   9. key escrow.

in the NS-004 pull request, the list was:

* certificate application;
* certificate validation;
* certificate approval;
* signing;
* certificate revocation;
* serving authoritative certificate status information; or
* key escrow.

I have resolved this conflict by taking "certificate validation" and
inserting it into the list as of NS-003 as the fifth item.

    -- tobij

* Remove "certificate validation" from the Certificate System definition

In the Cloud Services call, we agreed that the term is problematic as
"certificate validation" is something that is performed by browsers when
encountering a certificate on a website, and also otherwise not required
as it should be covered by the "identity validation" list item.

* Update NSR.md

Added a very minor addition to the Network Equipment definition by adding an "other" in "Hardware devices and OTHER components that..."

* Update NSR.md

Added some language for clarity to the Certificate Systems definition.

* Update NSR.md

Removed the word `minimally` from section 4 as it is already covered in the Introduction section. 

Made some verbiage changes to make the requirements clearer in section 4

* Reorder Section 4 and update from NSWG discussion

Based on discussion August 13, 2024, update Section 4.
Also reordering 4 of the 5 subsections to be more "sequential"

* Address comments

* Add effective date for CP/CPS SLO

* Update NSR.md

Added a preamble to section 4.4 that states: "The CA's vulnerability identification process MUST include monitoring for relevant security advisories and penetration testing."

* Update NSR.md

Added preamble for section 4.4 

"The CA's vulnerability identification process MUST include monitoring for relevant security advisories and penetration testing."

* Update NSR.md

Closing out some comments. Changes made:

1. Removed duplicate sentence about monitoring external sources for vulns

2. Updated language in 4.3

* Updates from 10 Sep 24 NSWG

Based on Dan's comments and discussion in 10 Sep 24 NSWG meeting, incorporating these changes prior to discussion period start.

* Update NSR.md

Updating the effective date to April 22, 2025 and making it optional to adhere to these new requirements prior to April 22, 2025.

* Update NSR.md

Changed the language in the ##Requirements section to read: 

"Effective 2025-04-29, the CA shall adhere to these Requirements"

This represents an effective date that is 7 days after NS-005 ballot goes into effect.

* Update NSR.md

Updated ## Requirements language.

* Update NSR.md

Remove unused definitions and clarify what NCSSR version is in force.
    
This was made in response to Corey Bonnell's feedback at https://groups.google.com/a/groups.cabforum.org/g/netsec/c/k4T0UfTRBDI/m/6LBjnhWbAQAJ


The date in force was revised to reflect what was used in NS005.

* Update NSR.md

Remove the OWASP Top 10 definition.

I missed it in my previous commit.

---------

Co-authored-by: Clint Wilson <clintw@apple.com>
Co-authored-by: Tobias S. Josefowitz <tobij@opera.com>
Co-authored-by: Cade Cairns <cade@carbon.net>
Co-authored-by: Clint Wilson <clint@wilsonovi.com>