Address 2.2.6 feedback
A concern was raised related to the current wording of "remote connection[s] that enable.... access to CA Infrastructure" as allowing an interpretation to conclude that a customer accessing a CA's web portal may qualify as meeting the description, thus necessitating the customer be brought into scope of the bulleted requirements. This commit aims to remove the possibility of this interpretation by introducing a more precise term ("Privileged Access") for the type of access enabled in Section 2.2.6 which then necessitates the behavior defined therein.
This term is also used in the definition of "Workstation", as it was identified by the group as very applicable in that context.

Separately, I'll note that while a customer portal may be a system operated by a CA, and it's very likely to be an application system which *interfaces* with CA Infrastructure as defined in the NCSSRs, it's not inherently (and likely should _not_ be) the infrastructure itself. That application layer functions at a fundamentally different layer than the management and administration layers underlying the CA Infrastructure and Network Boundary Controls these requirements seek to ensure the security of. Hopefully this change makes that more clear, though that separation is important enough that it shouldn't hinge solely upon this change.