NS-008: Updates to CA Infrastructure Scope, Trusted Roles, Systems' Applicability, and various other improvements (#44)
* CA Infrastructure Scope update

Following discussions in the NSWG, this is an example version of the NCSSRs which removes several of the "Systems" from the scope of CA Infrastructure (and the document in general).
The changes are based on the assumed completion of IPR review for NS-004 and NS-006.

Additionally, fixed some minor issues arising from NS-004 and NS-006:
* Reversion of Workstation definition
* Orphaned text in 1.2.2
* Markdown formatting in 2.2.1.2

* Update from comments

* Incorporated suggestions from several comments
* Moved Requirements of 4.1 to 1 (and relabeled 4 as 4.1 so as to not renumber all of 4's subsections)
* Attempted to address concerns with change in CA Infrastructure scope by:
  * re-adding (slightly modified) Security Support System definition
  * changing Trusted Role to apply to those with Certificate Systems or Root CA Systems access

* Incorporate NSWG Discussions

Based on discussions in NSWG meetings over the last couple months, changes have been made to:

* switch 1.1.1.1 and 1.1.1.2 ordering ✅
* update defined term used in Section 1 to CA Infrastructure ✅
* make SDN in 1.1.1.1 a MUST ✅
* require physically separate networks for Root CA Systems and 3rd Party ✅
* Replace MPC in 2.1 with "separation of duties" ✅
* consistency of Trusted Roles (Certificate Systems and Root CA Systems) throughout section 2 (CA Infrastructure and Network Equipment) ✅
* update formatting of Critical Security Event (i.e. make into list, fix 'or', etc.) ✅
* use of "and/or" vs "or" vs "and" in Section 2, review and ensure it makes sense ✅
* formatting bug in 2.2.1.2 ✅
* Consider extending 1.2.2 to include Network Equipment (at least in the SHOULD for within CA infra; maybe for a later version) ✅
* Fix risk assessment capitalization in 4.4 ✅
* Access control systems (and other parts of Security Support Systems) should be in-scope of Section 4's vulnerability management ✅ <- added future effective date

* Formatting

Fix a few instances of formatting inconsistency

* F2F Discussion updates

Based on discussion at CA/BF F2F 64:
* updated definition of Physically Secure Environment
* removed "network boundary control" from definition of Security Support System (now its own definition)
* removed "accounts within CA Infrastructure" from Multi-Factor Authentication requirements of 2.2.3
* added Network Boundary Control to scope of Section 4's SHOULD and futre MUST requirements

* Update workflow file

Based on https://github.com/actions/runner-images/issues/11101, updating the build-draft-docs.yml to use ubuntu-24.04 instead of ubuntu-20.04

* NSWG April 8, 2025 Updates

Based on discussion, updated 3 areas

* Address 2 comments

Update Document History Adopted and Effective Dates to be consistently formatted
Update Section 4, shifting 4.1 to 4, and 4.2 - 4.4 to 4.1 - 4.3 respectively.

* Fix link

* Remove critical from Separation of Duties

* Push effective date to November

and fix spelling error in Risk Assessment definition

* Address 2.2.6 feedback

A concern was raised related to the current wording of "remote connection[s] that enable.... access to CA Infrastructure" as allowing an interpretation to conclude that a customer accessing a CA's web portal may qualify as meeting the description, thus necessitating the customer be brought into scope of the bulleted requirements. This commit aims to remove the possibility of this interpretation by introducing a more precise term ("Privileged Access") for the type of access enabled in Section 2.2.6 which then necessitates the behavior defined therein.
This term is also used in the definition of "Workstation", as it was identified by the group as very applicable in that context.

Separately, I'll note that while a customer portal may be a system operated by a CA, and it's very likely to be an application system which *interfaces* with CA Infrastructure as defined in the NCSSRs, it's not inherently (and likely should _not_ be) the infrastructure itself. That application layer functions at a fundamentally different layer than the management and administration layers underlying the CA Infrastructure and Network Boundary Controls these requirements seek to ensure the security of. Hopefully this change makes that more clear, though that separation is important enough that it shouldn't hinge solely upon this change.

* Add Document Version

Add Document Version, Date, and Change History entry

* Post-ballot adoption and IPR Review Adminstrative Edits

Update dates following ballot voting and IPR review